requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
cloudipsp==1.0.2
Django==3.2.7
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
psycopg2-binary==2.9.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
stripe==2.60.0
urllib3==1.26.6
psycopg2==2.8.6
whitenoise==5.3.0
dj_database_url==0.5.0
